requirements.txt
async-timeout==4.0.3
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
mfrc522==0.0.7
numpy==2.0.0
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.7
requests==2.32.3
RPi.GPIO==0.7.1
six==1.16.0
spidev==3.6
spotipy==2.24.0
tzdata==2024.1
urllib3==2.2.2
